[filesystem] implement 'create' argument for _NOFS.movedir
in pyfilesystem2, movedir copies the content from one directory to
antoher, optionally creating the destination if create=True.

When the destination is exists and is a directory, shutil.move moves
the source directory inside the destinatinn directory.

shutil.copytree also fails if the destination already exists; so here
we resort to distutils.dir_util.copy_tree (it's part of python std lib)