[svgPath] add copy of parser module from svg.path package
The upstream code comes from here:

https://github.com/regebro/svg.path

I prefer to 'vendor' the svg path parser instead of adding it as a
requirement to fonttools, because this way we can patch it to call a
pen object directly instead of returning a list of Path instances.

Also I'm only interested in the parser module, the rest of svg.path
is not useful here.

The license of svg.path is MIT, so it's compatible with fonttools
license (BSD 2-clause).

I added the original code so it's clear from the subsequent commits
what changes are being applied.