[otBase] Make __hash__ and __eq__ shallow
Since we merge duplicate objects, there's no reason to recursively hash and compare
subordinates.

Note that this currently makes some fonts bigger because of the way extension
subtables are handled.  For example, IranNastaliq2.ttf grows a few kb's with
this commit.  That will go away when we stop compartmentalizing duplicate-merging
across Extension sublookups.