[otBase] Rewrite table packing
This moves table packing into fontTools.ttLib.tables.distillery
Currently overflows are NOT handled, but the algorithm is changed
from DFS to Dijkstra's algorithm.  This already handles more input
without overflow, including the input in the issue below:

https://github.com/fonttools/fonttools/issues/1105

The trick now is to find input that does overflow, so I can go ahead
and implement overflow resolution.

Also, this introduces a dependency on fibonacci_heap_mod.  To be
removed by copying a Fibonacci Heap implementation in-tree, which
will also ensure reproducible output.