rewrite plistlib to use lxml to parse/write
The module is now independent from the built-in plistlib, though it
is largely based on that.
It uses lxml to parse/write property list, instead of the expat parser
and the pure-python xml writer that plistlib uses.

The load/loads functions pass a custom PlistTarget object to the parser
that creates python data structures instead of Element objects.
The dump/dumps functions build an ElementTree then serialize it with
lxml.

The API is the same as the python3's plistlib (but only XML plist is
supported).

Two additional functions fromtree/totree are exported, which are used
for loading and dumping plist data embedded in GLIF xml tree:

- `fromtree` loads a plist from a previosly parsed element tree.
- `totree` does the reverse, and builds a new element tree from some
  plist-compatible data structure (dict, list, etc.).