don't enforce default layer to be first, etc.
follow the spec and set the default to the layer whose directory is
named "glyphs" (the name of the layer can be user-defined, else it
must be "public.default" -- though we don't validate that here yet).

Also, added a new `getGlyphSets` method to the reader that returns
an ordered dict of GlyphSets keyed by name;

and also a `defaultLayerName` property to LayerSet.

added some backward-compatibility methods from the old ufoLib api.