layerSet: add 'load' class constructor; remove defaultLayerName attrib
The alternative class constructor 'load' takes a reader instance and
returns a new LayerSet instance by loading the glyph sets from the
reader and passing the correct default layer name. This will be used
when loading the Font from a path.

However, it can be useful to construct a LayerSet from code, from an
interable of Layer objects (which may in turn be created from code and
not read from GlyphSet paths).