misc: make Transformation a namedtuple; use unbound readf/writef/deletef methods
defcon uses plain tuples for these, and client code is expecting to
index into it as if it were a tuple.

for the DataStore, we need to call the class attribute, otherwise self
is passed as first argument, but we want the reader/writer object instead

also 'fileNames' for the keys is what client code expect.