warn if passed an explicit 'structure' argument with dest being an FS object
it doesn't make sense to say structure='zip' if we are saving to an already instantiated
FS object. The 'structure' keyword argument only makes sense when the destination is
a path string, where we have a choice whether to create either a UFOZ (zipfile) or a
normal UFO package.
For an already instantiated FS object, we always write it out as a flat package,
without the root contents subfolder.