[varLib.cff] Fix issues in building cff2 variable font
in varLib.cff2: shouldn't iterate through a dict while deleting keys
in cffLib.specializer.py: when doing peephole optimization, both the first and second args may be lists. Test for both cases.

It does bother me to have this test for list type args in specializeCommands(), since it is needed only for the one case where CFF2CharStringMergePen is being used to build up a list of source master coordinates in place of a single  coordinate.  I would rather add a function argument "doPeephole=True" for specializeCommands(), to allow suppressing this test. It will be  rare that it would be possible to satisfy the conditions for doing this optimization in a variable font, and suppressing the test makes more sense to me than complicating the common case for the sake of a special case.