instancer: must redo iup_delta_optimize if the default coordinates have changed
If we modify the default instance coordinates, then the inferred deltas that
are left in gvar are no longer valid, so we need to calculate them using the
original default coordinates.
They are then re-optimized using the modified default coordinates.

Also, the default deltas returned from instantiateTupleVariationStore are now
already rounded to integer.