varLib: initialize all fields to None in new empty GDEF
code elsewhere assumes that all optional fields in OT tables are initialized
to None (that is the case when decompiling from a file).
This patch makes sure that the new GDEF table build by varLib when creating
a new VF is properly initialised.
Ideally we wouldn't have to do that manually, but the constructor would
take care of that. But otData-generated classes are special...