Convert OTF font files in test data to ttx files.
This required a fix to fontTools/cffLib. When reading a CFF2 variable font (VF) from XML, the VF state in FontDict and PrivateDict gets set correctly only if the VarStore element is read in before the FDArray. Fixed by adding code to force the VarStore element to be read in before the FDArray.

When debugging, the inconsistency in how the VF state in FontDict and PrivateDict was set bothered me, so I converted these in getters and setters. This required minor touches to other files.